refactor(html): tokenization state: *attribute*
