+++ html/parser/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "resworb-html-parser"
-version = "0.1.116"
+version = "0.1.117"
 license-file = "../../LICENSE"
 edition = "2021"
